requirements.txt
numpy==2.1.3
pandas==2.2.3
plum-dispatch==1.7.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
Spire.Doc==12.7.1
tzdata==2024.2
